Add the OpenIOC 1.0 to 1.1 script as a entrypoint script to ioc_writer.
